BigDecimal, Global Refactor - Add BigDecimal support - Remove sql libs from public - Setup dev and preview servers - Update build logic:   - k2ts-service is now library   - start for reactive development & preview for testing - Update cache logic:   - Now for navigation events (open new page) it's online first and then we use cache.

+++ frontend/shared/interop-sample/data/src/commonMain/kotlin/interopSample/cloud/dtos/WeatherDTO.kt
@@ -2,6 +2,7 @@ package interopSample.cloud.dtos
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
+import utils.BigDecimal
 
 @Serializable
 internal data class WeatherDTO(
@@ -13,7 +14,7 @@ internal data class WeatherDTO(
     @Serializable
     internal data class CurrentWeatherDTO(
         @SerialName("temperature")
-        val temperature: Double
+        val temperature: BigDecimal,
     )
 
     @Serializable
